% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_audience.R
\name{audienceaveRage}
\alias{audienceaveRage}
\title{hashtagfrequency}
\usage{
audienceaveRage(user, token = NULL)
}
\arguments{
\item{token}{Your Twitter token}

\item{utilisateur}{User ID or screen name}
}
\value{
list followers
}
\description{
hashtagfrequency
}
