% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_audience.R, R/user_in_depth.R
\name{\%>\%}
\alias{\%>\%}
\alias{audienceaveRage}
\alias{\%>\%}
\alias{user_in_depth}
\title{Average audience}
\usage{
audienceaveRage(user, token = NULL)

user_in_depth(user, token = NULL, n_tweets = NULL, output = NULL)
}
\arguments{
\item{token}{Your Twitter token}

\item{utilisateur}{User ID or screen name}

\item{utilisateur}{User ID or screen name}

\item{token}{Your Twitter token}
}
\value{
list followers

list followers
}
\description{
Average audience

Average audience
}
